backend/pyproject.toml
[tool.poetry]
name = "alfred"
version = "0.1.0"
description = "A personal feed app"
authors = ["Campus Hack <your.email@example.com>"]
readme = "README.md"
packages = [{include = "alfred", from = "src"}]
classifiers = ["Private :: Do Not Upload"]

[tool.poetry.dependencies]
python = "^3.11"
toml = "^0.10.2"
python-dotenv = "^1.0.0"
pydantic = "^1.10.7"
pyyaml = "^6.0"
pandas = "^1.5.3"
pyspark = "^3.4.0"
mlflow = "^2.4.2"
ipython = "^8.13.2"
great-expectations = "^0.16.11"
sqlalchemy = "^2.0.17"
fastapi = "^0.99.1"
uvicorn = "^0.22.0"
httpx = "^0.24.1"




[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
mypy = "^1.1.1"
flake8 = "^6.0.0"
pre-commit = "^3.2.0"
flake8-docstrings = "^1.7.0"
pydocstyle = "^6.3.0"
isort = "^5.12.0"
types-pyyaml = "^6.0.12.8"
sqlalchemy-stubs = "^0.4"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.4.2"
mkdocs-material = "^9.1.4"
mkdocstrings = "^0.20.0"
mkdocstrings-python = "^0.8.3"
pygments = "^2.15.0"
mkdocs-gitbook = "^0.0.1"
mkdocs-dracula-theme = "^1.0.4"

[tool.poetry.group.test.dependencies]
pytest = "^7.2.2"
pytest-cov = "^4.0.0"
great-expectations = "^0.16.7"


[tool.poetry.group.viz.dependencies]
streamlit = "^1.20.0"
plotly = "^5.14.0"
seaborn = "^0.12.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

#set isort profile to "black" and apply same formatting rules as black does, such as line length and commas.
[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

ignore_missing_imports = true

[tool.flake8]
exclude = ".venv"
ignore = ["E203", "E266", "E501", "W503", "F403", "F401"]
max-line-length = 80
max-complexity = 18
select = ["B", "C", "E", "F", "W", "T4", "B9"]
docstring-convention = "google"
